win and end battle works
